build: Add a makefile for Linux kernel.
Build kernel as part of the default target.

Add a dev target as an alias for droidcore.

system.img must have a dependency on kernel module builds,
as kernel modules get installed at lib/modules on the device.

A shell target must not share a name with the host side executable.

Introduce mp and mmp shell commands as parallel build shorthands.

Whitespace and syntax fixes.  Changes to variable naming.

Bug 760963.

Change-Id: I0c7e261f7376ecd5dd6e017ac756dc5f6d847697
Reviewed-on: http://git-master/r/24278
Tested-by: Oskari Jaaskelainen <oskarij@nvidia.com>
Reviewed-by: Jan-Petter Vainionpaa <jvainionpaa@nvidia.com>
Reviewed-by: Frank Thomas Bourgeois <fbourgeois@nvidia.com>