build: kernel: Support for Device Tree
This commit supports building the DTB from the DTS in the kernel and
installing it to a special dtb partition to be picked up by the bootloader.

Once everything has transistioned over to device tree, we'll be able to
remove most of this mess. Some of it will go away once we've got good
enough bootloader support that we don't need to append it to the image.

In the BoardConfig.mk for each board, it will need to specify:

    TARGET_USE_DTB := true
    TARGET_KERNEL_DT_NAME := tegra-ventana

and to test non-appended DTBs (handled by the bootloader):

    BOOTLOADER_SUPPORTS_DTB := true

This commit should have no change to kernels that don't support device
tree.

Change-Id: I1d9ba324b7967e28ba874ec795cc637d3ee21f4c
Reviewed-on: http://git-master/r/56432
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Tested-by: Nitin Kumbhar <nkumbhar@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Dan Willemsen <dwillemsen@nvidia.com>