build: fix stat path for Darwin build
One of the Android build system macros tries to conjure the
appropriate syntax for the flavor of stat expected to be present on
the host system. In the meantime, the kernel build depends on the GNU
version of stat being in the path.

So far, the standard fix for this has been to add GNU stat to the path
before BSD stat.  However, this has recently generated the expected
conflict: the kernel must have GNU stat show up first in the path, but
the rest of the build system must have BSD stat be first in the path
on Darwin.

This change adds GNU stat to the top of the path for the kernel build
only.

This creates a dependency on GNU_COREUTILS being set, which is new.
For that reason, this change also adds a check for that variable on
Darwin, as well as a check to ensure that 'stat' from GNU coreutils
is actually installed.

Change-Id: I80a14408bc3945a3b74fc2990602ae922e1d9065
Reviewed-on: http://git-master/r/61961
Reviewed-by: Atte Peltomaki <apeltomaki@nvidia.com>
Tested-by: Atte Peltomaki <apeltomaki@nvidia.com>
Reviewed-by: Frank Bourgeois <fbourgeois@nvidia.com>
Reviewed-by: Oskari Jaaskelainen <oskarij@nvidia.com>