Pass NVUB defines to nvmake submake
This passes the NVUB defines to the submake into the
GL driver.  That code is unlikely to be released as source.
However, it includes header files which contain
protection, so it needs to know about the NVUB defines.
bug 965015

Change-Id: I3b58781994da009097cc423c50e6b3f8b34cac01
Reviewed-on: http://git-master/r/96811
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Acorn Pooley <apooley@nvidia.com>
Tested-by: Acorn Pooley <apooley@nvidia.com>
Reviewed-by: Mark Stadler <mastadler@nvidia.com>
Original: bc04e8c87e7f9316bd2098c2eb577ac06d3ce51a