nvmake: Set NV_MANGLE_GLSI=0 variable in nvmake
When invoking nvmake, set the NV_MANGLE_GLSI
top-level flag to 0 to indicate that the nvglsi/
nvidia-glsi module's external symbols should not
be mangled (i.e., obfuscated) when building EGL
and GLES components.  The default behavior of
nvmake is to mangle symbols of libraries that do
not expose public APIs, which includes nvglsi.
However, not all clients of the nvglsi API are
built using nvmake in the mobile build systems,
and the mobile build systems currently don't have
support for nvmake-style symbol mangling.

bug 1332417

Change-Id: Ica1d5d755810dda480b4ac44e1aa8dda62fddb93
Reviewed-on: http://git-master/r/259286
Reviewed-by: Antoine Chauveau <achauveau@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Andy Ritger <aritger@nvidia.com>
Reviewed-by: Oskari Jaaskelainen <oskarij@nvidia.com>
Tested-by: Oskari Jaaskelainen <oskarij@nvidia.com>