nvmake: Use Android toolchain, not its P4 copy
Use the standard Android toolchain within the nvmake
build system. This ensures that Google updates to the
toolchain are automatically picked up.

To avoid breaking incremental builds, this change detects
if binaries built with the old toolchain are present in
the tree and clobbers them.

Bug 872779

Change-Id: I02da9a10ae658702f4c0004f6ba83b4a5551a147
Reviewed-on: http://git-master/r/278944
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Tested-by: Antoine Chauveau <achauveau@nvidia.com>
Reviewed-by: Jan-Petter Vainionpaa <jvainionpaa@nvidia.com>
Reviewed-by: Stefan Becker <stefanb@nvidia.com>