build: fix UBM flagging
After a merge the NVUB_SUPPORTS_ for the merged chips must be defined
and default to 1 until all conditionals have been removed from the
code. This is needed to support those repositories which can't use the
sifter due to a different delivery process.

Introduce NVUB_SUPPORTS_FLAG_LIST to remove code duplication when UBM
flags need to be exported to another make instance.

Introduce ubm_enable.mk to set NVUB_UNIFIED_BRANCHING_ENABLED to 1.
The contents of this file will be truncated by the sifter.

Bug 1357361

Change-Id: I05d0fd4fc1d2f09d4caa1c309c548221697ebea6
Reviewed-on: http://git-master/r/280705
Reviewed-by: Antoine Chauveau <achauveau@nvidia.com>
Reviewed-by: Stefan Becker <stefanb@nvidia.com>
Tested-by: Stefan Becker <stefanb@nvidia.com>