multilib: Switch exports to using generated sources
We have multiple intermediate directories when building multilib, so
assuming there is only one doesn't work. local-generated-sources-dir has
one directory, and copies it's contents to both of the intermediates.

In this case, we really just need a central place to store the file.
Adding it to LOCAL_ADDITIONAL_DEPENDENCIES does ensure that the dependency is
handled properly.

bug 1465011

Change-Id: I8352488b80da76f913652aa61af2642b15e58140
Reviewed-on: http://git-master/r/363745
Tested-by: Dan Willemsen <dwillemsen@nvidia.com>
Reviewed-by: Michael I Gold <gold@nvidia.com>
Reviewed-by: Dan Willemsen <dwillemsen@nvidia.com>