build: remove redundancy in bootwrapper build
EMMC_BOOT is no longer a valid option. no need to run bootwrapper
build twice. In addition, no need to fork a shell to run it.

if full_filesystem doesn't exist, do not build bootwrapper. this
implies the BOARD_SUPPORT_SIMULATION is not defined, so no need
to build binaries for pre-silicon support.

Change-Id: If49bf8bf8d571ead1a58611421ec6e1bca3260b2
Reviewed-on: http://git-master/r/403020
Reviewed-by: Riham Haidar <rhaidar@nvidia.com>
Tested-by: Riham Haidar <rhaidar@nvidia.com>