flash.sh: remove pre-typed "auto"
Removes one of the most annoying functionality - pretyped "auto". "auto"
is still the default mode when Enter is pressed.

Another improvement made is to remember the choice selected
previously. So if the user enters "tn8-p1761-a04", it will be the
default option to choose on the next run.

Bug 1521072

Change-Id: I493060543d1b902e849e21ffcc23ea66268875bd
Reviewed-on: http://git-master/r/423442
Tested-by: Q-Ha Park <qpark@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Mitch Luban <mluban@nvidia.com>